build from source produces binary in source dir
On my machine `make build` produces a binary in the source directory, there is no `./bin` to be found.  I'm on a Linux machine so I updated that section; I also updated the Mac section on the assumption that it behaves similarly, but feel free to tell me if I'm mistaken.  I can't seem to find any explicit reference in the Makefile that dictates where the binary ends up, so perhaps it's an artifact of my particular go/cmake/make versions/config.